intro/learn ai/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.4
certifi==2025.4.26
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.3.0
groq==0.26.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pydantic==2.11.5
pydantic_core==2.33.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dotenv==1.1.0
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
typing-inspection==0.4.1
typing_extensions==4.13.2
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
